int32 to int8

+++ examples/empty/src/Controller.cpp
@@ -1,19 +1,19 @@
 # include "Controller.hpp"
 
-Array<int32> getWherePush()
+Array<int8> getWherePush()
 {
-    Array<int32> result = {-1, -1};
-    int32 size = System::EnumerateGamepads().size();
+    Array<int8> result = {-1, -1};
+    int8 size = System::EnumerateGamepads().size();
     ClearPrint();
     
-    for(int32 i : step(size))
+    for(int8 i : step(size))
     {
         if (const auto gamepad = Gamepad(i))
         {
             for (auto [j, button] : Indexed(gamepad.buttons))
             {
                 if(button.pressed()){
-                    result = {i, (int32)j};
+                    result = {i, (int8)j};
                     break;
                 }
             }
